use bigint for all keys

+++ ddl-tpch/bin_flat/alltables.sql
@@ -4,8 +4,8 @@ use ${DB};
 drop table if exists lineitem;
 create external table lineitem 
 (L_ORDERKEY BIGINT,
- L_PARTKEY INT,
- L_SUPPKEY INT,
+ L_PARTKEY BIGINT,
+ L_SUPPKEY BIGINT,
  L_LINENUMBER INT,
  L_QUANTITY DOUBLE,
  L_EXTENDEDPRICE DOUBLE,
@@ -23,7 +23,7 @@ ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION '${LOCATION}/lineitem';
 
 drop table if exists part;
-create external table part (P_PARTKEY INT,
+create external table part (P_PARTKEY BIGINT,
  P_NAME STRING,
  P_MFGR STRING,
  P_BRAND STRING,
@@ -36,10 +36,10 @@ ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION '${LOCATION}/part/';
 
 drop table if exists supplier;
-create external table supplier (S_SUPPKEY INT,
+create external table supplier (S_SUPPKEY BIGINT,
  S_NAME STRING,
  S_ADDRESS STRING,
- S_NATIONKEY INT,
+ S_NATIONKEY BIGINT,
  S_PHONE STRING,
  S_ACCTBAL DOUBLE,
  S_COMMENT STRING) 
@@ -47,8 +47,8 @@ ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION '${LOCATION}/supplier/';
 
 drop table if exists partsupp;
-create external table partsupp (PS_PARTKEY INT,
- PS_SUPPKEY INT,
+create external table partsupp (PS_PARTKEY BIGINT,
+ PS_SUPPKEY BIGINT,
  PS_AVAILQTY INT,
  PS_SUPPLYCOST DOUBLE,
  PS_COMMENT STRING)
@@ -56,25 +56,25 @@ ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION'${LOCATION}/partsupp';
 
 drop table if exists nation;
-create external table nation (N_NATIONKEY INT,
+create external table nation (N_NATIONKEY BIGINT,
  N_NAME STRING,
- N_REGIONKEY INT,
+ N_REGIONKEY BIGINT,
  N_COMMENT STRING)
 ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION '${LOCATION}/nation';
 
 drop table if exists region;
-create external table region (R_REGIONKEY INT,
+create external table region (R_REGIONKEY BIGINT,
  R_NAME STRING,
  R_COMMENT STRING)
 ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION '${LOCATION}/region';
 
 drop table if exists customer;
-create external table customer (C_CUSTKEY INT,
+create external table customer (C_CUSTKEY BIGINT,
  C_NAME STRING,
  C_ADDRESS STRING,
- C_NATIONKEY INT,
+ C_NATIONKEY BIGINT,
  C_PHONE STRING,
  C_ACCTBAL DOUBLE,
  C_MKTSEGMENT STRING,
@@ -83,7 +83,7 @@ ROW FORMAT DELIMITED FIELDS TERMINATED BY '|' STORED AS TEXTFILE
 LOCATION '${LOCATION}/customer';
 
 drop table if exists orders;
-create external table orders (O_ORDERKEY INT,
+create external table orders (O_ORDERKEY BIGINT,
  O_CUSTKEY BIGINT,
  O_ORDERSTATUS STRING,
  O_TOTALPRICE DOUBLE,
